feat(NA): support performant cancelable behaviour on Windows
