+++ bazel/bazel_windows.go
@@ -23,7 +23,26 @@ import (
 	"os/exec"
 	"strings"
 	"syscall"
+
+	"github.com/bazelbuild/bazel-watcher/ibazel/log"
 )
+
+// Internal function useful to send CTRL+C events
+// to a given process on Windows OS
+func sendCtrlBreak(pid int) {
+	kernel32DLL, dllLoadingError := syscall.LoadDLL("kernel32.dll")
+	if dllLoadingError != nil {
+		log.Errorf("Error loading a required dll `kernel32.dll` when `IBAZEL_ABORT_COMPILATION_EARLY=1`: %v\n", dllLoadingError)
+	}
+	GenerateConsoleCtrlEventProc, findProcError := kernel32DLL.FindProc("GenerateConsoleCtrlEvent")
+	if findProcError != nil {
+		log.Errorf("Error calling FindProc with GenerateConsoleCtrlEvent: %v\n", findProcError)
+	}
+	ConsoleCtrlEventResult, _, eventError := GenerateConsoleCtrlEventProc.Call(syscall.CTRL_BREAK_EVENT, uintptr(pid))
+	if ConsoleCtrlEventResult == 0 {
+		log.Errorf("Error while generating a ConsoleCtrlEvent on pid %d: %v\n", pid, eventError)
+	}
+}
 
 func (b *bazel) newCommand(command string, args ...string) (*bytes.Buffer, *bytes.Buffer) {
 	b.ctx, b.cancel = context.WithCancel(context.Background())
@@ -55,7 +74,10 @@ func (b *bazel) newCommand(command string, args ...string) (*bytes.Buffer, *byte
 	// OS other than Windows which is the reason why newCommand fn was created both
 	// for Windows and Unix
 	bazelPath = fmt.Sprintf("\"%s\"", bazelPath)
-	b.cmd.SysProcAttr = &syscall.SysProcAttr{}
+	// Flag to always run this cmd as a new windows process group
+	// otherwise we cannot correctly send the ctrl+c event on
+	// canceleable methods
+	b.cmd.SysProcAttr = &syscall.SysProcAttr{CreationFlags: syscall.CREATE_NEW_PROCESS_GROUP}
 	b.cmd.SysProcAttr.CmdLine = fmt.Sprintf("%s %s", bazelPath, strings.Join(args[:], " "))
 
 	stdoutBuffer := new(bytes.Buffer)
@@ -72,4 +94,48 @@ func (b *bazel) newCommand(command string, args ...string) (*bytes.Buffer, *byte
 	}
 
 	return stdoutBuffer, stderrBuffer
+}
+
+func (b *bazel) BuildCancelable(cancelCh chan bool, args ...string) (*bytes.Buffer, error) {
+	stdoutBuffer, stderrBuffer := b.newCommand("build", append(b.args, args...)...)
+	doneCh := make(chan error)
+
+	go func() {
+		doneCh <- b.cmd.Run()
+	}()
+
+	select {
+	case e := <-doneCh:
+		_, _ = stdoutBuffer.Write(stderrBuffer.Bytes())
+
+		return stdoutBuffer, e
+	case <-cancelCh:
+		sendCtrlBreak(b.cmd.Process.Pid)
+		b.Cancel()
+		<-doneCh
+
+		return nil, nil
+	}
+}
+
+func (b *bazel) TestCancelable(cancelCh chan bool, args ...string) (*bytes.Buffer, error) {
+	stdoutBuffer, stderrBuffer := b.newCommand("test", append(b.args, args...)...)
+	doneCh := make(chan error)
+
+	go func() {
+		doneCh <- b.cmd.Run()
+	}()
+
+	select {
+	case e := <-doneCh:
+		_, _ = stdoutBuffer.Write(stderrBuffer.Bytes())
+
+		return stdoutBuffer, e
+	case <-cancelCh:
+		sendCtrlBreak(b.cmd.Process.Pid)
+		b.Cancel()
+		<-doneCh
+
+		return nil, nil
+	}
 }
